Add more documentation in README, and rename an interface
