+++ projects/fhi-angular-highcharts/package.json
@@ -4,7 +4,8 @@
   "peerDependencies": {
     "@angular/common": "^14.2.0",
     "@angular/core": "^14.2.0",
-    "@folkehelseinstituttet/style": "^4.0.0",
+    "@folkehelseinstituttet/style": "^4.2.0",
+    "bootstrap": "^5.2.0",
     "date-fns": "~2.29.0",
     "highcharts": "~10.3.0",
     "highcharts-angular": "~3.0.0"
