+++ projects/fhi-angular-highcharts/src/public-api.ts
@@ -5,5 +5,7 @@
 export * from './lib/fhi-angular-highcharts.service';
 export * from './lib/fhi-angular-highcharts.component';
 export * from './lib/fhi-angular-highcharts.module';
-export { DiagramOptions } from './lib/fhi-highcharts-config.model';
+
+export { DiagramSerie } from './lib/diagram-options.model';
+export { DiagramOptions } from './lib/diagram-options.model';
 export { DiagramTypes } from './lib/fhi-diagram-types/fhi-diagram-types';
